poplar: enable CONFIG_SUPPORT_EMMC_BOOT
It enables CONFIG_SUPPORT_EMMC_BOOT to get additional features of the
eMMC boot partitions, which are useful to initialize the bare eMMC chip
from factory.  More specifically, we need to execute the following mmc
commands to get the eMMC chip on Poplar ready for booting image flashed
on it.

 mmc rst-function 0 1
 mmc partconf 0 0 7 0
 mmc bootbus 0 1 0 0

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>